pyproject.toml
[tool.poetry]
name = "mox001-webservice"
version = "0.1.0"
description = "Autherisation and authentification service"
authors = ["Yevgeniy Kapustin <kapustaevg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
fastapi-cache2 = "^0.2.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
gunicorn = "^23.0.0"
pyjwt = "^2.9.0"
loguru = "^0.7.2"
pydantic-settings = "^2.6.0"
redis = "^5.1.1"
passlib = "^1.7.4"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
